requirements.txt
django>=4.2
psycopg2-binary
python-decouple
gunicorn
whitenoise
requests
